% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{baselineModel}
\alias{baselineModel}
\title{Baseline Model}
\usage{
baselineModel(useDT, meterDT, base.length = 12,
  date.format = "\%m/\%d/\%Y", gboost = FALSE, padding = 0)
}
\description{
Baseline Model
}
