% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebase_form.R
\name{ebase_form}
\alias{ebase_form}
\title{Format ebase output}
\usage{
ebase_form(out, dat, interval, maxinterp = 43200/interval)
}
\arguments{
\item{out}{data.frame for model output}

\item{dat}{data.frame as returned by \code{\link{ebase_prep}}}

\item{maxinterp}{}
}
\value{
Formatted output for \code{\link{ebase}} with interpolated rows removed as defined by \code{maxinterp}
}
\description{
Format ebase output
}
\details{
This function is used internally with \code{\link{ebase}} and should not be called by itself.
}
\examples{
library(dplyr)
library(lubridate)
library(doParallel)

# get four days of data
dat <- exdat \%>\% 
  filter(month(exdat$DateTimeStamp) == 6 & day(exdat$DateTimeStamp) \%in\% 1:4)

##
# run ebase with defaults, parallel

\dontrun{
# setup parallel backend
cl <- makeCluster(2)
registerDoParallel(cl)

out <- ebase(dat, interval = 900, H = 1.85, progress = TRUE, n.chains = 2)

stopCluster(cl)

dat <- ebase_prep(dat, H = 1.85, interval = 900, ndays = 1)

ebase_form(out, dat, interval = 900, maxinterp = 43200 / interval) 
}
}
