% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebase_plot.R
\name{ebase_plot}
\alias{ebase_plot}
\title{Plot results from EBASE}
\usage{
ebase_plot(res, instantaneous = TRUE)
}
\arguments{
\item{res}{output data frame from \code{\link{ebase}}}

\item{instantaneous}{logical indicating if results are instantaneous (default) or averaged to daily}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot results from EBASE
}
\details{
All metabolic estimates are plotted as positive values (\code{D} is represented as net ingassing).
}
\examples{
# plot instantaneous
ebase_plot(exres)

# plot daily-averaged
ebase_plot(exres, instantaneous = FALSE)
}
