% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebase_prep.R
\name{ebase_prep}
\alias{ebase_prep}
\title{Prepare data for ebase}
\usage{
ebase_prep(dat, H)
}
\arguments{
\item{dat}{input data frame}

\item{H}{numeric as single value for water column depth (m) or vector equal in length to number of rows in \code{dat}}
}
\value{
A data frame with additional columns required for \code{\link{ebase}}
}
\description{
Prepare data for ebase
}
\details{
Checks if all columns are present by matching those in \code{\link{exdat}}, converts dissolved oxygen from mg/L to mmol/m3, calculates the Schmidt number (unitless) from water temp (C) and salinity (psu), and calculates dissolved oxygen equilibrium concentration (mmol/m3) from salinity and temperature
}
\examples{
dat <- ebase_prep(exdat, H = 1.85)
head(dat)
}
