% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebase_schmidt.R
\name{ebase_schmidt}
\alias{ebase_schmidt}
\title{Schmidt number calculation}
\usage{
ebase_schmidt(temp, salt)
}
\arguments{
\item{temp}{numeric for temperature (C)}

\item{salt}{numeric for salinity (PSU)}
}
\value{
sc (unitless)
}
\description{
Schmidt number calculation
}
\details{
The Schmidt number is calculated for the air-sea gas transfer velocity.
}
\examples{
temp <- c(10, 20, 30)
salt <- c(30, 35, 40)
ebase_schmidt(temp = temp, salt = salt)
}
