% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{ebGetStation}
\alias{ebGetStation}
\title{Get Nearest Weather Station}
\usage{
ebGetStation(x, y, start.date = NULL, end.date = NULL,
  include = NULL, exclude = NULL)
}
\description{
Get Nearest Weather Station
}
