% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ebMeterFormat}
\alias{ebMeterFormat}
\title{Meter Format}
\usage{
ebMeterFormat(meter, useDT, meterDT, base.length, date.format, padding,
  base.min, perf.min, perf.length, interval)
}
\description{
Meter Format
}
