% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ebModifyData}
\alias{ebModifyData}
\title{Modify formatted data}
\usage{
ebModifyData(dataList, FUN, meterSubset = NULL, periods = c("baseline",
  "blackout", "performance"))
}
\description{
Modify formatted data
}
