% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{ebRawConvert}
\alias{ebRawConvert}
\title{Convert Raw Utility Data to Long-form}
\usage{
ebRawConvert(dat, utility = "sdge", id = c("account", "meter"))
}
\description{
Convert Raw Utility Data to Long-form
}
