% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings.R
\name{ebSavings}
\alias{ebSavings}
\title{Format Savings from Summary Object}
\usage{
ebSavings(x, deemed, groups = NULL, sites_subset = NULL)
}
\description{
Format Savings from Summary Object
}
\arguments{
\item{x}{Output from the \code{ebSummary} object.}
\item{deemed}{A named character vector reporting deemed savings estimates for each site.}
\item{groups}{(optional)A named character vector assigning sites to groups. If supplied, savings the output
is aggregated at the group-level}
\item{sites_subset}{(optional) A subset of the }
}
\value{A table that compares gross NMEC and deemed savings at the site- or group-level. The output can
be visualized using the \code{ebPlot} function.}