% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ebSummary.prediction}
\alias{ebSummary.prediction}
\title{Prediction and Savings Summary}
\usage{
\method{ebSummary}{prediction}(x, norm_modelList = NULL)
}
\description{
Prediction and Savings Summary
}
