% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdat.R
\docType{data}
\name{exdat}
\alias{exdat}
\title{Sample data from Apalachicola NERRS}
\format{
A \code{data.frame} object with 27648 rows and 6 columns
\describe{
\item{DateTimeStamp}{date and time, America/Jamaica time zone, 15 minute time step}
\item{DO_obs}{dissolved oxygen, mg/L}
\item{Temp}{water temperature, C}
\item{Sal}{salinity, psu}
\item{PAR}{total PAR, W/m2}
\item{WSpd}{num, m/s}
}
}
\usage{
exdat
}
\description{
Sample data from Apalachicola NERRS
}
\examples{
head(exdat)
}
\seealso{
Other utilities: 
\code{\link{exres}}
}
\concept{utilities}
\keyword{datasets}
