% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_plot.R
\name{fit_plot}
\alias{fit_plot}
\title{Plot observed and modeled dissolved oxygen}
\usage{
fit_plot(res)
}
\arguments{
\item{res}{output data frame from \code{\link{ebase}}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot observed and modeled dissolved oxygen
}
\details{
Dissolved oxygen (mmol/m3) is plotted as observed from the input data (points) and modelled (lines) based on inputs to \code{\link{ebase}}
}
\examples{

library(dplyr)
library(lubridate)
library(doParallel)

# get four days of data
dat <- exdat \%>\% 
  filter(month(exdat$DateTimeStamp) == 6 & day(exdat$DateTimeStamp) \%in\% 1:4)

##
# run ebase with defaults, parallel

# setup parallel backend
cl <- makeCluster(2)
registerDoParallel(cl)

res <- ebase(dat, interval = 900, H = 1.85, progress = TRUE, n.chains = 2)

stopCluster(cl)

# plot observed and modeled DO
fit_plot(res)
}
