% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{formatData}
\alias{formatData}
\title{Format Data}
\usage{
formatData(dt, id.var, use.var, date.var, hour.var, temp.var,
  install.date.var, nra.start.var, nra.end.var, base.length, date.format)
}
\description{
Format Data
}
