% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metab_update.R
\name{metab_update}
\alias{metab_update}
\title{Update metabolism jags fit}
\usage{
metab_update(metabfit, update.chains, n.iter)
}
\arguments{
\item{metabfit}{initial jags metabolism output}

\item{update.chains}{logical to update, only if TRUE}

\item{n.iter}{number of iterations}
}
\value{
Updated jags metabolism output
}
\description{
Update metabolism jags fit
}
\details{
This function is used by \code{\link{ebase}} and is not to be called directly by the user.  It provides additional model iterations if convergence is not achieved.
}
