% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{baselineModel}
\alias{baselineModel}
\title{Baseline Model}
\usage{
baselineModel(dt, id.var, use.var, date.var, hour.var, temp.var,
  install.date.var, nra.start.var = NULL, nra.end.var = NULL,
  base.length = NULL, date.format = NULL, yoy.adjust = NULL,
  include.gboost = NULL)
}
\description{
Baseline Model
}
