% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credible_plot.R
\name{credible_plot}
\alias{credible_plot}
\title{Plot credible intervals for a, r, and b}
\usage{
credible_plot(res, params = c("a", "r", "b"))
}
\arguments{
\item{res}{output data frame from \code{\link{ebase}}}

\item{params}{character vector indicating which parameters to plot, one to any of \code{a}, \code{r}, or \code{b},  (default all)}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot credible intervals for a, r, and b
}
\details{
This function plots 95\% credible intervals (2.5th to 97.5th percentiles, approximate posterior distributions) for \code{a}, \code{r} and/or \code{b} using the output from \code{\link{ebase}}. Results in the plot are grouped by the \code{ndays} argument that was used in \code{\link{ebase}}.
}
\examples{

library(dplyr)
library(lubridate)
library(doParallel)

# get four days of data
dat <- exdat \%>\% 
  filter(month(exdat$DateTimeStamp) == 6 & day(exdat$DateTimeStamp) \%in\% 1:4)

##
# run ebase with defaults, parallel

# setup parallel backend
cl <- makeCluster(2)
registerDoParallel(cl)

res <- ebase(dat, interval = 900, H = 1.85, progress = TRUE, n.chains = 2)

stopCluster(cl)

# plot credible intervals
credible_plot(res)
}
