% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credible_prep.R
\name{credible_prep}
\alias{credible_prep}
\title{Get credible intervals for a, r, b}
\usage{
credible_prep(res, params = c("a", "r", "b"), labels = FALSE)
}
\arguments{
\item{res}{output data frame from \code{\link{ebase}}}

\item{params}{character vector indicating which parameters to plot, one to any of \code{a}, \code{r}, or \code{b} (default all)}

\item{labels}{logical indicating of parameter labels are output as an expression for parsing in plot facets, default \code{FALSE}}
}
\value{
A data frame
}
\description{
Get credible intervals for a, r, b
}
\details{
This function gets 95\% credible intervals (2.5th to 97.5th percentiles, approximate posterior distributions) for \code{a}, \code{r}, and/or \code{b} using the output from \code{\link{ebase}}. The function is used in \code{\link{credible_plot}}, but is provided as a separate function for convenience.
}
\examples{

library(dplyr)
library(lubridate)
library(doParallel)

# get four days of data
dat <- exdat \%>\% 
  filter(month(exdat$DateTimeStamp) == 6 & day(exdat$DateTimeStamp) \%in\% 1:4)

##
# run ebase with defaults, parallel

# setup parallel backend
cl <- makeCluster(2)
registerDoParallel(cl)

res <- ebase(dat, interval = 900, H = 1.85, progress = TRUE, n.chains = 2)

stopCluster(cl)

# get credible intervals
credible_prep(res)
}
