% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebase.R
\name{ebase}
\alias{ebase}
\title{Estuarine Bayesian Single-station Estimation method for ecosystem metabolism}
\usage{
ebase(
  dat,
  H,
  interval,
  ndays = 1,
  arng = c(0, 2),
  rvar = 100,
  brng = c(0.8, 1.2),
  interp = TRUE,
  maxgap = 1e+06,
  inits = NULL,
  n.iter = 10000,
  update.chains = TRUE,
  n.burnin = n.iter * 0.5,
  n.chains = 3,
  n.thin = 10,
  progress = FALSE,
  model_file = NULL
)
}
\arguments{
\item{dat}{input data frame}

\item{H}{numeric as single value for water column depth (m) or vector equal in length to number of rows in \code{dat}}

\item{interval}{timestep interval in seconds}

\item{ndays}{numeric for number of days in \code{dat} for optimizing the metabolic equation, see details}

\item{arng}{numeric for proportional range of the prior distribution for the \emph{a} parameter, see details}

\item{rvar}{numeric for the variance estimate of the prior distribution for the \emph{r} parameter, see details}

\item{brng}{numeric for the proportional range of the prior distribution for the \emph{b} parameter, see details}

\item{interp}{logical indicating if linear interpolation is used to fill all data gaps, only applies if timestep specified by \code{interval} is not uniform}

\item{maxgap}{numeric vector indicating maximum gap size to interpolate where size is number of records at the value specified by \code{interval}, e.g., if \code{interval = 900} and \code{maxgap = 4}, gaps up to and including one hour will be linearly interpolated}

\item{inits}{\code{NULL} or a function that returns a named list of starting values for the parameters to be estimated in the jags model, see examples}

\item{n.iter}{number of MCMC iterations, passed to \code{\link[R2jags]{jags}}}

\item{update.chains}{logical to run \code{\link{metab_update}} if chains do not converge}

\item{n.burnin}{number of MCMC chains to delete, passed to \code{\link[R2jags]{jags}}}

\item{n.chains}{number of MCMC chains to run, passed to \code{\link[R2jags]{jags}}}

\item{n.thin}{number of nth iterations to save for each chain, passed to \code{\link[R2jags]{jags}}}

\item{progress}{logical if progress saved to a txt file named 'log.txt' in the working directory}

\item{model_file}{\code{NULL} to use the model file included with the package or a path to a model text file can be used}
}
\value{
A data frame with metabolic estimates for volumetric gross production (\code{Pg_vol}, O2 mmol/m3/d), respiration (\code{Rt_vol},  O2 mmol/m3/d), and gas exchange (\code{D}, O2 mmol/m3/d).  Additional parameters estimated by the model that are returned include \code{a} and \code{b}.  The \code{a} parameter is a constant that represents the primary production per quantum of light with units of  (mmol/m3/d)(W/m2) and is used to estimate gross production (Grace et al., 2015).  The \code{b} parameter is a constant used to estimate gas exchange in (cm/hr)/(m2/s2) (provided as 0.251 in eqn. 4 in Wanninkhof 2014).  Observed dissolved oxygen (\code{DO_obs}, mmol/m3), modeled dissolved oxygen (\code{DO_mod}, mmol/m3), and delta dissolved oxygen of the modeled results (\code{dDO}, mmol/m3/d) are also returned.  Note that delta dissolved oxygen is a daily rate.

Model fit can also be assessed using the \code{converge} and \code{rsq} columns.  The values in these columns apply to each group in the \code{grp} column as specified with the \code{ndays} argument. The \code{converge} column indicates \code{"Check convergence"} or \code{"Fine"} if the JAGS estimate converged at that iteration (repeated across rows for the group).  The \code{n.chains} argument can be increased if convergence is not achieved. Similarly, the \code{rsq} column shows the r-squared values of the linear fit between the modeled and observed dissolved oxygen (repeated across rows for the group).  These values can also be viewed with \code{\link{fit_plot}}.
}
\description{
Estuarine Bayesian Single-station Estimation method for ecosystem metabolism
}
\details{
Required input data are time series for dissolved oxygen (mg/L), water temperature (C), salinity (psu), total PAR (W/m2/s), and wind speed (m/s).  See the \code{\link{exdat}} example data file for a representation of the required data.  Data are typically from continuously monitored water quality and weather parameters are hourly of sub-hourly time steps.  Oxygen concentrations are converted to mmol/m3 prior to metabolic estimation. Water column depth is also required to return volumetric estimates.  This can be supplied as a single value or a vector of length equal to the number of rows in \code{dat}.

The metabolic estimates are based on a mass balance equation in Grace et al. 2015 with the gas exchange estimate from Wanninkhof 2004.  It is similar to that provided by the BASEmetab R package at \url{https://github.com/dgiling/BASEmetab}, with modifications to estimate different parameters. The equation optimized in the JAGS model is:

\deqn{ \frac{\delta{C_d}}{\delta{t}} = [\,aPAR]\, - [\,r]\, - \frac{1}{H}\left[-bU_{10}^2\left(\frac{s_c}{600} \right)^{-0.5} \left(C_{Sat} - C_d \right )\right]}

Gross production is provided by \emph{aPAR}, respiration is provided by \emph{r}, and gas exchange is provided by the remainder.  The likelihood of the parameters \emph{a}, \emph{r}, and \emph{b} given the observed data are estimated from the JAGS model using prior distributions shown in the model file. At each time step, the change in oxygen concentration between time steps is calculated from the equation using model inputs and parameter guesses, and then a finite difference approximation is used to estimate modeled oxygen concentration.  The estimated concentration is also returned for comparison to observed as one measure of model performance.

The prior distributions for the \emph{a}, \emph{r}, and \emph{b} parameters are defined in the model file included with the package as uninformed uniform, normal, and uniform distributions, respectively.  The prior distribution for \emph{a} as default varies uniformly across the range from zero to 200\\% of an estimated mean of 0.2 (mmol/m3/d)(W/m2).  This range can be changed using the \code{arng} parameter, default being 0 to 2 for zero to 200\\%.  The prior distribution for the \emph{r} parameter follows a normal distribution with mean zero and variance equal to 100, but truncated to not go below zero.  This variance estimate can be changed with the \code{rvar} parameter.  The prior distribution for \emph{b} as default varies uniformly across the range from 80 to 120\\% of an estimated mean of 0.251 (cm/hr)/(m2/s2).  This range can be changed using the \code{brng} parameter, default being 0.8 to 1.2 for 80 to 120\\%.  Note that Wanninkhof (2014) states that the uncertainty of the \emph{b} parameter typically does not vary more than 20\%, although this range is generally unknown in estuarine systems.

The \code{ndays} argument defines the number of days that are used for optimizing the above mass balance equation.  By default, this is done each day, i.e., \code{ndays= 1} such that a loop is used that applies the model equation to observations within each day, evaluated iteratively from the first observation in a day to the last.  Individual parameter estimates for \emph{a}, \emph{r}, and \emph{b} are then returned for each day.  However, more days can be used to estimate the unknown parameters, such that the loop can be evaluated for every \code{ndays} specified by the argument.  The \code{ndays} argument will separate the input data into groups of consecutive days, where each group has a total number of days equal to \code{ndays}.  The final block may not include the complete number of days specified by \code{ndays} if the number of unique dates in the input data includes a remainder when divided by \code{ndays}, e.g., if seven days are in the input data and \code{ndays = 5}, there will be two groups where the first has five days and the second has two days. The output data from \code{ebase} includes a column that specifies the grouping that was used based on \code{ndays}.

Missing values in the input data can also be interpolated prior to estimating metabolism.  This is the default behavior and it is the responsibility of the user to verify that these interpolated values are not wildly inaccurate.  Missing values are linearly interpolated between non-missing values at the time step specified by the value in \code{interval}.  This works well for small gaps, but can easily create inaccurate values at gaps larger than a few hours. The \code{\link{interp_plot}} function can be used to visually assess the interpolated values.
}
\examples{

library(dplyr)
library(lubridate)
library(doParallel)

# get four days of data
dat <- exdat \%>\% 
  filter(month(exdat$DateTimeStamp) == 6 & day(exdat$DateTimeStamp) \%in\% 1:4)

##
# run ebase with defaults, parallel

# setup parallel backend
cl <- makeCluster(2)
registerDoParallel(cl)

res <- ebase(dat, interval = 900, H = 1.85, progress = TRUE, n.chains = 2)

stopCluster(cl)

##
# run ebase with different initial starting values for the three parameters, parallel

inits <- function(interval = 900){
  list(
    a = 0.2 / (86400 / interval),
    r = 20 / (86400 / interval),
    b = 0.251 / (100 * 3600 / interval)
  )
}

# setup parallel backend
cl <- makeCluster(2)
registerDoParallel(cl)
  
res <- ebase(dat, interval = 900, H = 1.85, progress = TRUE, inits = inits, n.chains = 2)

stopCluster(cl)
}
\references{
Grace, M.R., Giling, D.P., Hladyz, S., Caron, V., Thompson, R.M., Nally, R.M., 2015. Fast processing of diel oxygen curves: Estimating stream metabolism with BASE (BAyesian Single-station Estimation). Limnology and Oceanography: Methods 13, e10011. https://doi.org/10.1002/lom3.10011

Wanninkhof, R., 2014. Relationship between wind speed and gas exchange over the ocean revisited. Limnology and Oceanography: Methods 12, 351–362. https://doi.org/10.4319/lom.2014.12.351
}
