% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebase.R
\name{ebase}
\alias{ebase}
\title{Estuarine Bayesian Single-station Estimation method for ecosystem metabolism}
\usage{
ebase(
  dat,
  H,
  interval,
  inits = NULL,
  n.iter = 10000,
  update.chains = TRUE,
  n.burnin = n.iter * 0.5,
  n.chains = 3,
  n.thin = 10,
  progress = FALSE
)
}
\arguments{
\item{dat}{input data frame}

\item{H}{numeric as single value for water column depth (m) or vector equal in length to number of rows in \code{dat}}

\item{interval}{timestep interval in seconds}

\item{inits}{\code{NULL} or a function that returns a named list of starting values for the parameters to be estimated in the jags model, see examples}

\item{n.iter}{number of MCMC iterations, passed to \code{\link[R2jags]{jags}}}

\item{update.chains}{logical to run \code{\link{metab_update}} if chains do not converge}

\item{n.burnin}{number of MCMC chains to delete, passed to \code{\link[R2jags]{jags}}}

\item{n.chains}{number of MCMC chains to run, passed to \code{\link[R2jags]{jags}}}

\item{n.thin}{number of nth iterations to save for each chain, passed to \code{\link[R2jags]{jags}}}

\item{progress}{logical if progress saved to a txt file names 'log.txt' in the working directory}
}
\description{
Estuarine Bayesian Single-station Estimation method for ecosystem metabolism
}
\examples{
\dontrun{

library(dplyr)
library(lubridate)
library(doParallel)

# get four days of data
dat <- exdat \%>\% 
  filter(month(exdat$DateTimeStamp) == 6 & day(exdat$DateTimeStamp) \%in\% 1:4)

##
# run ebase with defaults

# setup parallel backend
ncores <- detectCores()
cl <- makeCluster(ncores - 2)
registerDoParallel(cl)
  
res <- ebase(dat, interval = 900, H = 1.85, progress = T)

stopCluster(cl)

##
# run ebase with different initial starting values for the three parameters

inits <- function(){
  list(
    a = 0.2 / troc,
    r = 20 / troc,
    b = 0.251 / 400
  )
}

# setup parallel backend
ncores <- detectCores()
cl <- makeCluster(ncores - 2)
registerDoParallel(cl)
  
res <- ebase(dat, interval = 900, H = 1.85, progress = T, inits = inits)

stopCluster(cl)

}
}
