% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebase_plot.R
\name{ebase_plot}
\alias{ebase_plot}
\title{Plot results from EBASE}
\usage{
ebase_plot(res, instantaneous = TRUE)
}
\arguments{
\item{res}{output data frame from \code{\link{ebase}}}

\item{instantaneous}{logical indicating if results are instantaneous (default) or averaged to daily}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot results from EBASE
}
\details{
All metabolic estimates are plotted as positive values (\code{D} is represented as net ingassing).
}
\examples{
library(dplyr)
library(lubridate)
library(doParallel)

# get four days of data
dat <- exdat \%>\% 
  filter(month(exdat$DateTimeStamp) == 6 & day(exdat$DateTimeStamp) \%in\% 1:4)

# run ebase with defaults, parallel

# setup parallel backend
cl <- makeCluster(2)
registerDoParallel(cl)

res <- ebase(dat, interval = 900, Z = 1.85, progress = TRUE, n.chains = 2)

stopCluster(cl)

# plot instantaneous
ebase_plot(res)

# plot daily-averaged
ebase_plot(res, instantaneous = FALSE)
}
