% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ebDataFormat}
\alias{ebDataFormat}
\title{Data Format}
\usage{
ebDataFormat(useDT, meterDT, base.length = 365, perf.length = 365,
  interval = c("daily", "hourly"), padding = 0, base.min = 0,
  perf.min = 0)
}
\description{
Data Format
}
