% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ebDataFormat}
\alias{ebDataFormat}
\title{Data Format}
\usage{
ebDataFormat(x, install_dates = NULL, sites = NULL, norm_data = NULL,
  data_options = NULL, temp_bins = NULL, cpus = 1)
}
\description{
Data Format
}
